+++ src/migration/triple-store-user-configuration-migration.js
